Introduce DTOs, query builders, and authentication scaffolding
- Added DTO classes using `spatie/laravel-data` to standardize
  application data transfer and request handling.
- Introduced reusable query builder abstractions for filtering,
  searching, sorting, eager loading, and pagination logic.
- Added transaction and category query implementations to encapsulate
  domain-specific query behavior.
- Implemented login/logout authentication flow with session
  regeneration and protected dashboard routing.
- Added typed request validation with `LoginRequest` and `LoginData`
  integration.
- Added controller scaffolding for balances, budgets, categories,
  dashboard, and transactions to prepare feature modules.
- Updated models with enum and integer casts to improve type safety and
  ensure consistent monetary value handling.
- Changed migration monetary columns from `decimal` to `bigInteger`
  for integer-based financial storage consistency.
- Added initial authenticated and guest route separation.

+++ database/migrations/2026_05_12_050849_create_budgets_table.php
@@ -18,9 +18,9 @@ return new class extends Migration
             $table->foreignIdFor(User::class)->constrained();
             $table->foreignIdFor(Category::class)->constrained();
             $table->string('type');
-            $table->decimal('planned_amount')->default(0);
-            $table->decimal('actual_amount')->default(0);
-            $table->decimal('diff_amount')->default(0);
+            $table->bigInteger('planned_amount')->default(0);
+            $table->bigInteger('actual_amount')->default(0);
+            $table->bigInteger('diff_amount')->default(0);
             $table->timestamps();
         });
     }
