+++ database/migrations/2026_05_12_050853_create_transactions_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->foreignIdFor(Category::class)->constrained();
             $table->string('type');
             $table->date('date');
-            $table->decimal('amount')->default(0);
+            $table->bigInteger('amount')->default(0);
             $table->text('description')->nullable();
             $table->timestamps();
         });
